backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.26.0
chromadb>=0.4.22
email-validator>=2.1.0
docker>=7.1.0
slowapi==0.1.9
numpy<2.0.0
sentence-transformers>=2.2.2
pypdf>=4.0.0
google-auth
urllib3>=2.0.0
python-docx>=1.1.0
tiktoken>=0.6.0
faster-whisper>=0.10.0
soundfile>=0.12.1
pydub>=0.25.1
beautifulsoup4>=4.12.0
Pillow>=10.0.0
edge-tts>=6.1.9
